refactor : 인테셉터 리팩토링
- 공통된 기능은 상위로 다른 기능은 하위 구현체로 뺐다
