--- src/main/java/nextstep/auth/authentication/TokenAuthenticationInterceptor.java
@@ -1,50 +1,0 @@
-package nextstep.auth.authentication;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import nextstep.auth.authentication.convertor.TokenConvertor;
-import nextstep.auth.context.Authentication;
-import nextstep.auth.token.JwtTokenProvider;
-import nextstep.auth.token.TokenResponse;
-import nextstep.member.application.CustomUserDetailsService;
-import nextstep.member.domain.LoginMember;
-import org.springframework.http.MediaType;
-import org.springframework.web.servlet.HandlerInterceptor;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class TokenAuthenticationInterceptor implements HandlerInterceptor {
-
-    private CustomUserDetailsService customUserDetailsService;
-    private JwtTokenProvider jwtTokenProvider;
-    private ObjectMapper objectMapper = new ObjectMapper();
-    private TokenConvertor tokenConvertor;
-
-    public TokenAuthenticationInterceptor(CustomUserDetailsService customUserDetailsService, JwtTokenProvider jwtTokenProvider, TokenConvertor convert) {
-        this.customUserDetailsService = customUserDetailsService;
-        this.jwtTokenProvider = jwtTokenProvider;
-        this.tokenConvertor = convert;
-    }
-
-    @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
-        AuthenticationToken authenticationToken = tokenConvertor.convert(request);
-        Authentication authentication = authenticate(authenticationToken);
-        String payload = objectMapper.writeValueAsString(authentication.getPrincipal());
-        TokenResponse tokenResponse = new TokenResponse(jwtTokenProvider.createToken(payload));
-
-        String responseToClient = new ObjectMapper().writeValueAsString(tokenResponse);
-        response.setStatus(HttpServletResponse.SC_OK);
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.getOutputStream().print(responseToClient);
-
-        return false;
-    }
-
-    public Authentication authenticate(AuthenticationToken authenticationToken) {
-        // TODO: AuthenticationToken에서 AuthenticationToken 객체 생성하기
-        LoginMember loginMember = customUserDetailsService.loadUserByUsername(authenticationToken.getPrincipal());
-        return new Authentication(loginMember);
-    }
-}
