+++ src/main/java/nextstep/auth/authentication/interceptor/SessionAuthenticationInterceptor.java
@@ -1,0 +1,28 @@
+package nextstep.auth.authentication.interceptor;
+
+import nextstep.auth.authentication.convertor.AuthenticationConverter;
+import nextstep.auth.context.Authentication;
+import nextstep.auth.context.SecurityContext;
+import nextstep.member.application.CustomUserDetailsService;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+import static nextstep.auth.context.SecurityContextHolder.SPRING_SECURITY_CONTEXT_KEY;
+
+public class SessionAuthenticationInterceptor extends AuthenticationInterceptor {
+
+
+    public SessionAuthenticationInterceptor(CustomUserDetailsService userDetailsService, AuthenticationConverter converter) {
+        super(userDetailsService, converter);
+    }
+
+    @Override
+    public void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) {
+        HttpSession httpSession = request.getSession();
+        httpSession.setAttribute(SPRING_SECURITY_CONTEXT_KEY, new SecurityContext(authentication));
+        response.setStatus(HttpServletResponse.SC_OK);
+    }
+
+}
