+++ src/test/java/nextstep/subway/unit/TokenAuthenticationInterceptorTest.java
@@ -2,12 +2,14 @@ package nextstep.subway.unit;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import nextstep.auth.authentication.AuthenticationToken;
-import nextstep.auth.authentication.TokenAuthenticationInterceptor;
 import nextstep.auth.authentication.convertor.TokenConvertor;
+import nextstep.auth.authentication.interceptor.TokenAuthenticationInterceptor;
 import nextstep.auth.context.Authentication;
 import nextstep.auth.token.JwtTokenProvider;
 import nextstep.auth.token.TokenRequest;
 import nextstep.member.application.CustomUserDetailsService;
+import nextstep.member.domain.LoginMember;
+import nextstep.member.domain.Member;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -59,6 +61,7 @@ class TokenAuthenticationInterceptorTest {
         //given
         MockHttpServletRequest mockRequest = createMockRequest();
         AuthenticationToken authenticationToken = tokenConvertor.convert(mockRequest);
+        when(customUserDetailsService.loadUserByUsername(anyString())).thenReturn(LoginMember.of(new Member(EMAIL, PASSWORD, 3)));
         //when
         Authentication authenticate = interceptor.authenticate(authenticationToken);
         //then
@@ -70,6 +73,7 @@ class TokenAuthenticationInterceptorTest {
         //given
         MockHttpServletRequest mockRequest = createMockRequest();
         MockHttpServletResponse mockResponse = new MockHttpServletResponse();
+        when(customUserDetailsService.loadUserByUsername(anyString())).thenReturn(LoginMember.of(new Member(EMAIL, PASSWORD, 3)));
 
         when(jwtTokenProvider.createToken(anyString())).thenReturn(JWT_TOKEN);
         //when
